+++ scout/server/blueprints/cases/views.py
@@ -26,6 +26,7 @@ from flask_weasyprint import HTML, render_pdf
 from scout.constants import CUSTOM_CASE_REPORTS, SAMPLE_SOURCE
 from scout.server.extensions import gens, mail, matchmaker, rerunner, store
 from scout.server.utils import (
+    html_2_pdf_file,
     institute_and_case,
     jsonconverter,
     templated,
@@ -739,27 +740,25 @@ def coverage_qc_report(institute_id, case_name):
 
     coverage_qc_report = data["case"]["coverage_qc_report"]
     report_format = request.args.get("format", "html")
-    if report_format == "pdf":
-        try:  # file could not be available
-            html_file = open(coverage_qc_report, "r")
-            source_code = html_file.read()
-            return render_pdf(
-                HTML(string=source_code),
-                download_filename=case_obj["display_name"]
-                + "_"
-                + datetime.datetime.now().strftime("%Y-%m-%d")
-                + "_coverage_qc_report.pdf",
-            )
-        except Exception as ex:
-            flash(
-                "An error occurred while downloading delivery report {} -- {}".format(
-                    coverage_qc_report, ex
-                ),
-                "warning",
-            )
 
-    out_dir = os.path.dirname(coverage_qc_report)
+    out_dir = os.path.abspath(os.path.dirname(coverage_qc_report))
     filename = os.path.basename(coverage_qc_report)
+
+    if report_format == "pdf":
+        bytes_file = html_2_pdf_file(coverage_qc_report, "landscape", 1000)
+        file_name = "_".join(
+            [
+                case_obj["display_name"],
+                datetime.datetime.now().strftime("%Y-%m-%d"),
+                "coverage_qc_report.pdf",
+            ]
+        )
+        return send_file(
+            bytes_file,
+            attachment_filename=file_name,
+            mimetype="application/pdf",
+            as_attachment=True,
+        )
 
     return send_from_directory(out_dir, filename)